% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_standard_curve.R
\name{add_curve_statistical_annotations}
\alias{add_curve_statistical_annotations}
\title{Add statistical annotations to summary table}
\usage{
add_curve_statistical_annotations(
  summary_table,
  statistical_results,
  reference_group
)
}
\description{
Add statistical annotations to summary table
}
\keyword{internal}
