% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_delta_delta_ct.R
\name{add_statistical_annotations}
\alias{add_statistical_annotations}
\title{Add statistical annotations to summary table}
\usage{
add_statistical_annotations(summary_table, statistical_results)
}
\description{
Add statistical annotations to summary table
}
\keyword{internal}
