% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_standard_curve.R
\name{calc_standard_curve}
\alias{calc_standard_curve}
\title{Standard Curve Calculation for qPCR}
\usage{
calc_standard_curve(
  cq_table,
  concentration_table,
  highest_concentration,
  lowest_concentration,
  dilution_factor = 4,
  use_mean_cq = TRUE
)
}
\arguments{
\item{cq_table}{A data frame containing position and Cq values}

\item{concentration_table}{A data frame containing position and concentration values}

\item{highest_concentration}{The highest concentration for calculation}

\item{lowest_concentration}{The lowest concentration for calculation}

\item{dilution_factor}{Dilution factor of cDNA template (default: 4)}

\item{use_mean_cq}{Whether to calculate using mean Cq values (default: TRUE)}
}
\value{
A list containing:
  \item{table}{Data frame with regression statistics and efficiency}
  \item{figure}{ggplot object showing the standard curve}
}
\description{
This function calculates the standard curve and amplification efficiency
for qPCR primers. Based on the amplification efficiency, it determines
which method can be used to calculate expression levels.
}
\examples{
# Load example data
cq_data_path <- system.file("examples", "calsc.cq.txt", package = "qPCRtools")
concentration_data_path <- system.file("examples", "calsc.info.txt", package = "qPCRtools")
cq_data <- read.table(cq_data_path, header = TRUE)
concentration_data <- read.table(concentration_data_path, header = TRUE)

# Calculate standard curve
result <- calc_standard_curve(
  cq_table = cq_data,
  concentration_table = concentration_data,
  lowest_concentration = 4,
  highest_concentration = 4096,
  dilution_factor = 4,
  use_mean_cq = TRUE
)

# View results
result$table
result$figure

}
\author{
Xiang LI <lixiang117423@gmail.com>
}
