% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_qpcr_efficiency.R
\name{calculate_corrected_expression_original}
\alias{calculate_corrected_expression_original}
\title{Calculate corrected expression using original algorithm}
\usage{
calculate_corrected_expression_original(
  df_expression,
  reference_gene,
  df_factor
)
}
\description{
Calculate corrected expression using original algorithm
}
\keyword{internal}
