% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_standard_curve.R
\name{calculate_curve_expression}
\alias{calculate_curve_expression}
\title{Calculate expression using standard curves}
\usage{
calculate_curve_expression(cq_table, curve_table, design_table)
}
\description{
Calculate expression using standard curves
}
\keyword{internal}
