% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_delta_delta_ct.R
\name{calculate_ddct_expression}
\alias{calculate_ddct_expression}
\title{Calculate delta-delta Ct expression values}
\usage{
calculate_ddct_expression(merged_data, reference_gene, reference_group)
}
\description{
Calculate delta-delta Ct expression values
}
\keyword{internal}
