% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_qpcr_efficiency.R
\name{calculate_normalization_factors_original}
\alias{calculate_normalization_factors_original}
\title{Calculate normalization factors using original algorithm}
\usage{
calculate_normalization_factors_original(df_expression, reference_gene)
}
\description{
Calculate normalization factors using original algorithm
}
\keyword{internal}
