% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_delta_ct.R
\name{calculate_reference_means}
\alias{calculate_reference_means}
\title{Calculate reference gene mean Cq values}
\usage{
calculate_reference_means(merged_data, reference_gene)
}
\description{
Calculate reference gene mean Cq values
}
\keyword{internal}
