% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_delta_ct.R
\name{calculate_target_expression}
\alias{calculate_target_expression}
\title{Calculate target gene expression levels}
\usage{
calculate_target_expression(merged_data, reference_data, reference_gene)
}
\description{
Calculate target gene expression levels
}
\keyword{internal}
