% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_to_df.R
\name{colnames_to_df}
\alias{colnames_to_df}
\title{Convert Column Names to a Single-Column Data Frame}
\usage{
colnames_to_df(df, name = "name")
}
\arguments{
\item{df}{A data frame or matrix.}

\item{name}{Character string for the output column name (default: "name").}
}
\value{
A single-column data frame.
}
\description{
Extract column names of a data frame (or matrix) and return them as a
single-column data frame. Useful for building sample metadata tables from
the column names of a count matrix in a pipe chain:

\preformatted{
df \%>\%
  colnames_to_df("sample") \%>\%
  dplyr::filter(sample != "description") \%>\%
  dplyr::mutate(group = stringr::str_split(sample, "R") \%>\% sapply("[", 1))
}
}
\examples{
\dontrun{
df <- data.frame(T1R1 = 1:3, T1R2 = 4:6, T2R1 = 7:9)
colnames_to_df(df, "sample")

# Pipe into further wrangling
df \%>\%
  colnames_to_df("sample") \%>\%
  dplyr::mutate(group = stringr::str_split(sample, "R") \%>\% sapply("[", 1))
}

}
\author{
Xiang LI <lixiang117423@gmail.com>
}
