% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_standard_curve.R
\name{create_curve_expression_plot}
\alias{create_curve_expression_plot}
\title{Create standard curve expression plot}
\usage{
create_curve_expression_plot(summary_data, plot_type, plot_ncol)
}
\description{
Create standard curve expression plot
}
\keyword{internal}
