% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_delta_delta_ct.R
\name{create_ddct_plot}
\alias{create_ddct_plot}
\title{Create delta-delta Ct plot}
\usage{
create_ddct_plot(summary_data, plot_type, plot_ncol)
}
\description{
Create delta-delta Ct plot
}
\keyword{internal}
