% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_delta_ct.R
\name{create_expression_plot}
\alias{create_expression_plot}
\title{Create expression level plot}
\usage{
create_expression_plot(summary_table)
}
\description{
Create expression level plot
}
\keyword{internal}
