% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_qpcr_efficiency.R
\name{create_plot_original}
\alias{create_plot_original}
\title{Create plot using original approach}
\usage{
create_plot_original(res_all, plot_type, plot_ncol)
}
\description{
Create plot using original approach
}
\keyword{internal}
