% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_standard_curve.R
\name{create_standard_curve_plot}
\alias{create_standard_curve_plot}
\title{Create standard curve plot}
\usage{
create_standard_curve_plot(processed_data, use_mean_cq)
}
\description{
Create standard curve plot
}
\keyword{internal}
