% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.ideo.karyotype}
\alias{df.ideo.karyotype}
\title{Human karyotype (test data for plot_ideogram)}
\format{
## `df.ideo.karyotype`
A data frame with 24 rows and 5 columns:
\describe{
  \item{Chr}{chromosome identifier}
  \item{Start}{chromosome start (bp)}
  \item{End}{chromosome length / end (bp)}
  \item{CE_start}{centromere start (bp)}
  \item{CE_end}{centromere end (bp)}
}
}
\source{
RIdeogram (Hao et al. 2020, PeerJ Comput. Sci. 6:e251), Artistic-2.0.
}
\usage{
df.ideo.karyotype
}
\description{
Human chromosomes with centromere positions, from the RIdeogram package.
}
\keyword{datasets}
