% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.rnaseq.go}
\alias{df.rnaseq.go}
\title{Test data.}
\format{
## `df.rnaseq.go`
A data frame with 6 rows and 1 columns:
\describe{
  \item{gene}{gene name}
  \item{go_id}{GO term id}
  \item{go_term}{GO term name}
  \item{go_ontology}{GO term ontology}
}
}
\source{
Xiang Li
}
\usage{
df.rnaseq.go
}
\description{
Test data.
}
\keyword{datasets}
