% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_qpcr_efficiency.R
\name{find_reference_genes_original}
\alias{find_reference_genes_original}
\title{Find reference genes using original GeNorm implementation}
\usage{
find_reference_genes_original(df_expression)
}
\description{
Find reference genes using original GeNorm implementation
}
\keyword{internal}
