% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_qpcr_efficiency.R
\name{merge_qpcr_data}
\alias{merge_qpcr_data}
\title{Merge data exactly as in original}
\usage{
merge_qpcr_data(cq_table, design_table)
}
\description{
Merge data exactly as in original
}
\keyword{internal}
