% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_standard_curve.R
\name{merge_standard_curve_data}
\alias{merge_standard_curve_data}
\title{Merge data and handle gene information flexibly}
\usage{
merge_standard_curve_data(cq_table, concentration_table, gene_name)
}
\description{
Merge data and handle gene information flexibly
}
\keyword{internal}
