% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_stat.R, R/scale01.R
\name{mutate_scale01}
\alias{mutate_scale01}
\title{Convenient function for scaling in dplyr pipelines}
\usage{
mutate_scale01(.data, ...)

mutate_scale01(.data, ...)
}
\arguments{
\item{.data}{A data frame (can be grouped)}

\item{...}{Column names to scale (unquoted)}
}
\value{
A data frame with additional scaled columns

A data frame with additional scaled columns
}
\description{
A wrapper function that makes it easy to scale multiple columns 
in dplyr pipelines with automatic "_scaled" suffix.

A wrapper function that makes it easy to scale multiple columns 
in dplyr pipelines with automatic "_scaled" suffix.
}
\examples{
library(dplyr)

df <- data.frame(
  group = rep(c("A", "B"), each = 3),
  var1 = c(1, 3, 5, 2, 4, 6),
  var2 = c(10, 30, 50, 20, 40, 60)
)

# Scale within groups
df \%>\%
  group_by(group) \%>\%
  mutate_scale01(var1, var2)

library(dplyr)

df <- data.frame(
  group = rep(c("A", "B"), each = 3),
  var1 = c(1, 3, 5, 2, 4, 6),
  var2 = c(10, 30, 50, 20, 40, 60)
)

# Scale within groups
df \%>\%
  group_by(group) \%>\%
  mutate_scale01(var1, var2)

}
\author{
Xiang LI \email{lixiang117423@foxmail.com}
}
