% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_microbiome_net.R
\name{net2ggnetview}
\alias{net2ggnetview}
\title{Convert microbiome_net Result to ggNetView tbl_graph}
\usage{
net2ggnetview(net_result, group)
}
\arguments{
\item{net_result}{Output from \code{\link{microbiome_net}}.}

\item{group}{Character string specifying which group to convert.}
}
\value{
A \code{tidygraph::tbl_graph} object with node attributes
  (name, Modularity, Degree, Strength) and edge attributes
  (correlation, weight, corr_direction).
}
\description{
Converts the output of \code{\link{microbiome_net}} for a specified group
into a \code{tbl_graph} object compatible with \code{ggNetView::ggNetView()}.
}
\examples{
\dontrun{
res <- microbiome_net(df.asv, df.sample, groupCol = "group")
tg <- net2ggnetview(res, group = "T1")
ggNetView::ggNetView(tg, layout = "gephi", add_outer = TRUE, label = TRUE)
}
}
\seealso{
\code{\link{microbiome_net}}
}
\author{
Xiang LI <lixiang117423@gmail.com>
}
