% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_standard_curve.R
\name{normalize_by_reference_gene}
\alias{normalize_by_reference_gene}
\title{Normalize expression by reference gene}
\usage{
normalize_by_reference_gene(expression_data, reference_gene)
}
\description{
Normalize expression by reference gene
}
\keyword{internal}
