% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_color_scales.R
\name{pal_sci}
\alias{pal_sci}
\alias{pal_nature}
\alias{pal_science}
\alias{pal_cell}
\alias{pal_jacs}
\alias{pal_fuel}
\alias{pal_chem_eng}
\alias{pal_nat_comm}
\alias{pal_shinkai}
\title{Academic journal color palettes}
\usage{
pal_sci(palette = c("default"), alpha = 1)

pal_nature(palette = c("default", "materials"), alpha = 1)

pal_science(palette = c("default"), alpha = 1)

pal_cell(palette = c("default"), alpha = 1)

pal_jacs(palette = c("default"), alpha = 1)

pal_fuel(palette = c("default"), alpha = 1)

pal_chem_eng(palette = c("default"), alpha = 1)

pal_nat_comm(palette = c("default"), alpha = 1)

pal_shinkai(palette = c("default", "blue_tones", "red_tones"), alpha = 1)
}
\arguments{
\item{palette}{Palette type. Available options depend on the journal:
\itemize{
  \item sci: "default" (12-color balanced scientific palette)
  \item nature: "default", "materials" (12-color Nature-inspired palettes)
  \item science: "default" (12-color Science journal palette)
  \item cell: "default" (12-color Cell journal palette)
  \item jacs: "default" (12-color JACS palette)
  \item fuel: "default" (10-color high-contrast palette)
  \item chem_eng: "default" (12-color Chemical Engineering Journal palette)
  \item nat_comm: "default" (12-color Nature Communications palette)
  \item shinkai: "default", "blue_tones", "red_tones" (Makoto Shinkai inspired)
}}

\item{alpha}{Transparency level, a real number in (0, 1].
See `alpha` in [grDevices::rgb()] for details.}
}
\description{
Color palettes inspired by plots in academic journals.
Designed for scientific data visualization with carefully selected colors
that work well for both print and digital formats.
}
\examples{
library("scales")
show_col(pal_sci("default")(10))
show_col(pal_sci("default", alpha = 0.6)(10))
}
\author{
Xiang LI \email{lixiang117423@foxmail.com}
}
