% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_standard_curve.R
\name{perform_curve_statistical_analysis}
\alias{perform_curve_statistical_analysis}
\title{Perform statistical analysis for standard curve data}
\usage{
perform_curve_statistical_analysis(
  expression_data,
  reference_group,
  statistical_method
)
}
\description{
Perform statistical analysis for standard curve data
}
\keyword{internal}
