% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_delta_delta_ct.R
\name{perform_statistical_analysis}
\alias{perform_statistical_analysis}
\title{Perform statistical analysis}
\usage{
perform_statistical_analysis(
  expression_data,
  reference_group,
  statistical_method
)
}
\description{
Perform statistical analysis
}
\keyword{internal}
