% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_qpcr_efficiency.R
\name{perform_statistical_analysis_original}
\alias{perform_statistical_analysis_original}
\title{Perform statistical analysis using original method}
\usage{
perform_statistical_analysis_original(
  res_all,
  reference_group,
  statistical_method
)
}
\description{
Perform statistical analysis using original method
}
\keyword{internal}
