% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bio.R
\name{+.theme_bio_list}
\alias{+.theme_bio_list}
\title{Add Operator for theme_bio_list}
\usage{
\method{+}{theme_bio_list}(e1, e2)
}
\arguments{
\item{e1}{A theme_bio_list object or theme}

\item{e2}{A theme_bio_list object or theme}
}
\value{
A combined theme_bio_list object
}
\description{
Custom + operator to combine two theme_bio_list objects or add
a theme to a theme_bio_list.
}
\keyword{internal}
