% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_delta_delta_ct.R
\name{prepare_ddct_data}
\alias{prepare_ddct_data}
\title{Prepare and merge data for delta-delta Ct calculation}
\usage{
prepare_ddct_data(cq_table, design_table)
}
\description{
Prepare and merge data for delta-delta Ct calculation
}
\keyword{internal}
