% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_delta_ct.R
\name{prepare_delta_ct_data}
\alias{prepare_delta_ct_data}
\title{Prepare and merge data for delta-Ct calculation}
\usage{
prepare_delta_ct_data(cq_table, design_table)
}
\description{
Prepare and merge data for delta-Ct calculation
}
\keyword{internal}
