% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_core_microbiome.R
\name{print.core_microbiome}
\alias{print.core_microbiome}
\title{Print Method for Core Microbiome Results}
\usage{
\method{print}{core_microbiome}(x, ...)
}
\description{
Print Method for Core Microbiome Results
}
\author{
Xiang LI \email{lixiang117423@foxmail.com}
}
