% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{qq_build_single}
\alias{qq_build_single}
\title{Build single-group QQ plot}
\usage{
qq_build_single(
  qq_df,
  line_df,
  title,
  x_lab,
  y_lab,
  point_color,
  point_size,
  point_alpha,
  ribbon_fill,
  ribbon_alpha,
  line_color,
  line_width,
  line_type,
  base_size
)
}
\description{
Build single-group QQ plot
}
\keyword{internal}
