% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pfam.R
\name{quick_pfam_plot}
\alias{quick_pfam_plot}
\title{Quick function for plotting protein domains}
\usage{
quick_pfam_plot(data, tree_file = NULL, ...)
}
\arguments{
\item{data}{Path to pfam file}

\item{tree_file}{Path to tree file (optional)}

\item{...}{Other parameters passed to plot_pfam}
}
\value{
ggplot object
}
\description{
Quick function for plotting protein domains
}
\examples{
\dontrun{
quick_pfam_plot("path/to/pfam.tsv")
quick_pfam_plot("path/to/pfam.tsv", tree_file = "path/to/tree.nwk")
}
}
\author{
Xiang LI <lixiang117423@gmail.com>
}
