% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_delta_delta_ct.R
\name{remove_expression_outliers}
\alias{remove_expression_outliers}
\title{Remove outliers using IQR method}
\usage{
remove_expression_outliers(expression_data)
}
\description{
Remove outliers using IQR method
}
\keyword{internal}
