% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cor_p.R
\name{safe_cor_test_p}
\alias{safe_cor_test_p}
\title{Compute correlation p-value, returning NA on failure}
\usage{
safe_cor_test_p(x, y, method)
}
\description{
Compute correlation p-value, returning NA on failure
}
\keyword{internal}
