% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_color_scales.R
\name{scale_color_sci}
\alias{scale_color_sci}
\alias{scale_colour_sci}
\alias{scale_fill_sci}
\alias{scale_color_nature}
\alias{scale_colour_nature}
\alias{scale_fill_nature}
\alias{scale_color_science}
\alias{scale_colour_science}
\alias{scale_fill_science}
\alias{scale_color_cell}
\alias{scale_colour_cell}
\alias{scale_fill_cell}
\alias{scale_color_jacs}
\alias{scale_colour_jacs}
\alias{scale_fill_jacs}
\alias{scale_color_fuel}
\alias{scale_colour_fuel}
\alias{scale_fill_fuel}
\alias{scale_color_chem_eng}
\alias{scale_colour_chem_eng}
\alias{scale_fill_chem_eng}
\alias{scale_color_nat_comm}
\alias{scale_colour_nat_comm}
\alias{scale_fill_nat_comm}
\alias{scale_color_shinkai}
\alias{scale_colour_shinkai}
\alias{scale_fill_shinkai}
\title{Academic journal color scales}
\usage{
scale_color_sci(palette = c("default"), alpha = 1, ...)

scale_colour_sci(palette = c("default"), alpha = 1, ...)

scale_fill_sci(palette = c("default"), alpha = 1, ...)

scale_color_nature(palette = c("default", "materials"), alpha = 1, ...)

scale_colour_nature(palette = c("default", "materials"), alpha = 1, ...)

scale_fill_nature(palette = c("default", "materials"), alpha = 1, ...)

scale_color_science(palette = c("default"), alpha = 1, ...)

scale_colour_science(palette = c("default"), alpha = 1, ...)

scale_fill_science(palette = c("default"), alpha = 1, ...)

scale_color_cell(palette = c("default"), alpha = 1, ...)

scale_colour_cell(palette = c("default"), alpha = 1, ...)

scale_fill_cell(palette = c("default"), alpha = 1, ...)

scale_color_jacs(palette = c("default"), alpha = 1, ...)

scale_colour_jacs(palette = c("default"), alpha = 1, ...)

scale_fill_jacs(palette = c("default"), alpha = 1, ...)

scale_color_fuel(palette = c("default"), alpha = 1, ...)

scale_colour_fuel(palette = c("default"), alpha = 1, ...)

scale_fill_fuel(palette = c("default"), alpha = 1, ...)

scale_color_chem_eng(palette = c("default"), alpha = 1, ...)

scale_colour_chem_eng(palette = c("default"), alpha = 1, ...)

scale_fill_chem_eng(palette = c("default"), alpha = 1, ...)

scale_color_nat_comm(palette = c("default"), alpha = 1, ...)

scale_colour_nat_comm(palette = c("default"), alpha = 1, ...)

scale_fill_nat_comm(palette = c("default"), alpha = 1, ...)

scale_color_shinkai(
  palette = c("default", "blue_tones", "red_tones"),
  alpha = 1,
  ...
)

scale_colour_shinkai(
  palette = c("default", "blue_tones", "red_tones"),
  alpha = 1,
  ...
)

scale_fill_shinkai(
  palette = c("default", "blue_tones", "red_tones"),
  alpha = 1,
  ...
)
}
\arguments{
\item{palette}{Palette type. Available options depend on the journal:
\itemize{
  \item sci: "default" (12-color balanced scientific palette)
  \item nature: "default", "materials" (12-color Nature-inspired palettes)
  \item science: "default" (12-color Science journal palette)
  \item cell: "default" (12-color Cell journal palette)
  \item jacs: "default" (12-color JACS palette)
  \item fuel: "default" (10-color high-contrast palette)
  \item chem_eng: "default" (12-color Chemical Engineering Journal palette)
  \item nat_comm: "default" (12-color Nature Communications palette)
  \item shinkai: "default", "blue_tones", "red_tones" (Makoto Shinkai inspired)
}}

\item{alpha}{Transparency level, a real number in (0, 1].
See `alpha` in [grDevices::rgb()] for details.}

\item{...}{Additional parameters for [ggplot2::discrete_scale()].}
}
\description{
See [pal_sci()] and related palette functions for details.
}
\examples{
library("ggplot2")
data("iris")

# Basic usage
ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width, color = Species)) +
  geom_point(size = 2) +
  scale_color_sci()

# With transparency
ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width, color = Species)) +
  geom_point(size = 2) +
  scale_color_sci(alpha = 0.7)
}
