% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wgcna.R
\name{summary.wgcna_results}
\alias{summary.wgcna_results}
\title{Summary method for WGCNA results}
\usage{
\method{summary}{wgcna_results}(object, ...)
}
\arguments{
\item{object}{A wgcna_results object}

\item{...}{Additional arguments (not used)}
}
\description{
Summary method for WGCNA results
}
