% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_standard_curve.R
\name{validate_curve_inputs}
\alias{validate_curve_inputs}
\title{Validate input parameters for standard curve calculation}
\usage{
validate_curve_inputs(
  cq_table,
  curve_table,
  design_table,
  reference_gene,
  reference_group,
  statistical_method,
  plot_type
)
}
\description{
Validate input parameters for standard curve calculation
}
\keyword{internal}
