% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_delta_delta_ct.R
\name{validate_ddct_inputs}
\alias{validate_ddct_inputs}
\title{Validate input parameters for delta-delta Ct calculation}
\usage{
validate_ddct_inputs(
  cq_table,
  design_table,
  reference_gene,
  reference_group,
  statistical_method,
  plot_type
)
}
\description{
Validate input parameters for delta-delta Ct calculation
}
\keyword{internal}
