% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_delta_ct.R
\name{validate_delta_ct_inputs}
\alias{validate_delta_ct_inputs}
\title{Validate input parameters for delta-Ct calculation}
\usage{
validate_delta_ct_inputs(cq_table, design_table, reference_gene)
}
\description{
Validate input parameters for delta-Ct calculation
}
\keyword{internal}
