% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expression_qpcr_efficiency.R
\name{validate_qpcr_inputs}
\alias{validate_qpcr_inputs}
\title{Validate input parameters - keeping original requirements}
\usage{
validate_qpcr_inputs(cq_table, design_table, statistical_method, plot_type)
}
\description{
Validate input parameters - keeping original requirements
}
\keyword{internal}
