% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_standard_curve.R
\name{calc_standard_curve}
\alias{calc_standard_curve}
\title{Standard Curve Calculation for qPCR}
\usage{
calc_standard_curve(
  cq_table,
  concentration_table,
  gene_name = NULL,
  highest_concentration,
  lowest_concentration,
  dilution_factor = 4,
  use_mean_cq = TRUE
)
}
\arguments{
\item{cq_table}{A data frame containing position and Cq values.
Must contain columns: Position, Cq}

\item{concentration_table}{A data frame containing position and concentration values.
Must contain columns: Position, Conc. May also contain Gene column.}

\item{gene_name}{Optional character string specifying gene name. If NULL and Gene
column exists in concentration_table, will use that. If NULL and no Gene column,
will use "Unknown_Gene" (default: NULL)}

\item{highest_concentration}{The highest concentration for calculation}

\item{lowest_concentration}{The lowest concentration for calculation}

\item{dilution_factor}{Dilution factor of cDNA template (default: 4)}

\item{use_mean_cq}{Whether to calculate using mean Cq values (default: TRUE)}
}
\value{
A list containing:
  \item{table}{Data frame with regression statistics and efficiency}
  \item{plot}{ggplot object showing the standard curve}
}
\description{
This function calculates the standard curve and amplification efficiency
for qPCR primers. Based on the amplification efficiency, it determines
which method can be used to calculate expression levels.
}
\examples{
# Load example data
cq_data_path <- system.file("extdata/qPCR", "calsc.cq.txt", package = "bioRtools")
concentration_data_path <- system.file("extdata/qPCR", "calsc.info.txt", package = "bioRtools")
cq_data <- read.table(cq_data_path, header = TRUE)
concentration_data <- read.table(concentration_data_path, header = TRUE)

# Calculate standard curve (gene info in concentration_table)
result1 <- calc_standard_curve(
  cq_table = cq_data,
  concentration_table = concentration_data,
  lowest_concentration = 4,
  highest_concentration = 4096,
  dilution_factor = 4,
  use_mean_cq = TRUE
)

# Calculate standard curve (specify gene name manually)
result2 <- calc_standard_curve(
  cq_table = cq_data,
  concentration_table = concentration_data,
  gene_name = "MyGene",
  lowest_concentration = 4,
  highest_concentration = 4096,
  dilution_factor = 4,
  use_mean_cq = TRUE
)

# View results
result1$table
result1$plot

}
\author{
Xiang LI <lixiang117423@gmail.com>
}
